+++ CHANGELOG.md
@@ -13,3 +13,4 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Project Boilerplate
 - Linter for enforcing code style (Google) and finding and fixing common problems
+- Added music player component and page